+++ sites/jeromefitzgerald.com/src/components/AppBar/MenuMobile.tsx
@@ -25,7 +25,7 @@ import NextLink from 'next/link'
 import * as React from 'react'
 import { useSwipeable } from 'react-swipeable'
 
-import useStore from '~store/useStore'
+import useAudio from '~store/useAudio'
 
 const slideIn = keyframes({
   from: { transform: '$$transformValue' },
@@ -56,7 +56,7 @@ const MenuMobile = ({ handleSelect, navigationNonMutated }) => {
    * @custom to sheet
    */
   const { theme } = useTheme()
-  const audio = useStore.use.audio()
+  const audio = useAudio.use.audio()
   const [open, openSet] = React.useState(false)
   const handleSelectInternal = (event, item) => {
     void handleSelect(event, item)
