+++ sites/jeromefitzgerald.com/src/components/Toggle/ToggleAudio.tsx
@@ -9,13 +9,13 @@ import { SpeakerOffIcon, SpeakerModerateIcon } from '@radix-ui/react-icons'
 import * as React from 'react'
 import { useSound } from 'use-sound'
 
-import useStore from '~store/useStore'
+import useAudio from '~store/useAudio'
 
 const ToggleAudio = (props) => {
-  const audio = useStore.use.audio()
-  const audioToggle = useStore.use.audioToggle()
-  const sounds = useStore.use.sounds()
-  const volume = useStore.use.volume()
+  const audio = useAudio.use.audio()
+  const audioToggle = useAudio.use.audioToggle()
+  const sounds = useAudio.use.sounds()
+  const volume = useAudio.use.volume()
   const [playEnableSound] = useSound(sounds.enableSound, {
     soundEnabled: true,
     volume,
