+++ sites/jeromefitzgerald.com/src/components/Toggle/ToggleTheme.tsx
@@ -11,15 +11,15 @@ import { useTheme } from 'next-themes'
 import * as React from 'react'
 import { useSound } from 'use-sound'
 
-import useStore from '~store/useStore'
+import useAudio from '~store/useAudio'
 
 const ThemeToggle = (props) => {
   const { theme, setTheme } = useTheme()
   const content = `Toggle theme to ${theme === 'light' ? 'dark' : 'light'}`
 
-  const audio = useStore.use.audio()
-  const sounds = useStore.use.sounds()
-  const volume = useStore.use.volume()
+  const audio = useAudio.use.audio()
+  const sounds = useAudio.use.sounds()
+  const volume = useAudio.use.volume()
 
   const [playBleep] = useSound(sounds.bleep, {
     soundEnabled: audio,
