+++ sites/jeromefitzgerald.com/src/lib/notion/app/utils/TextAnnotations.tsx
@@ -6,7 +6,7 @@ import * as React from 'react'
 import { useSound } from 'use-sound'
 
 import { nextSeo } from '~config/index'
-import useStore from '~store/useStore'
+import useAudio from '~store/useAudio'
 import getNextLink from '~utils/getNextLink'
 
 /**
@@ -25,9 +25,9 @@ const EmojiParser = dynamic(
 const domain = new URL(nextSeo.url)
 
 const TextAnnotationLink = ({ children, href }) => {
-  const audio = useStore.use.audio()
-  const sounds = useStore.use.sounds()
-  const volume = useStore.use.volume()
+  const audio = useAudio.use.audio()
+  const sounds = useAudio.use.sounds()
+  const volume = useAudio.use.volume()
   const [playPopDown] = useSound(sounds.popDown, {
     soundEnabled: audio,
     volume,
