+++ sites/jeromefitzgerald.com/src/store/slices/audio/index.ts
@@ -1,0 +1,2 @@
+export { Audio } from './audio'
+export type { IAudio } from './audio.types'
